✨ fix
第一遍跑通

+++ src/main/java/moe/kouyou/mikan/script/exec/ValueOperation.kt
@@ -1,10 +1,10 @@
 package moe.kouyou.mikan.script.exec
 
-inline operator fun MNumber.plus(right: MNumber) = MNumber(this.value + right.value)
-inline operator fun MNumber.minus(right: MNumber) = MNumber(this.value - right.value)
-inline operator fun MNumber.times(right: MNumber) = MNumber(this.value * right.value)
-inline operator fun MNumber.div(right: MNumber) = MNumber(this.value / right.value)
-inline operator fun MNumber.rem(right: MNumber) = MNumber(this.value % right.value)
+inline operator fun MNumber.plus(right: MNumber) = MNumber(this.value as Double + right.value as Double)
+inline operator fun MNumber.minus(right: MNumber) = MNumber(this.value as Double - right.value as Double)
+inline operator fun MNumber.times(right: MNumber) = MNumber(this.value as Double * right.value as Double)
+inline operator fun MNumber.div(right: MNumber) = MNumber(this.value as Double / right.value as Double)
+inline operator fun MNumber.rem(right: MNumber) = MNumber(this.value as Double % right.value as Double)
 
 /*
 inline operator fun MString.plus(right: MNumber) = MString(this.value + right.value)
